[APPC-1248] Refactorization: Gamification progress bar is now a custom view to avoid code replication
